Improving documents (#15)
* Rename NewDocumentCollection => NewDocumentsCollection
* Replace RootHash to Checksum according to changed terminology. Additional tests.
* Updated readme
