+++ README.md
@@ -1,3 +1,14 @@
-# sdk-go
+<div align="center">
+  <img class="logo" src="https://www.obada.io/assets/images/obada-logo.svg" alt="OBADA | Blockchain Node"/>
+</div>
 
-OBADA sdk implementation in go
+The “sdkgo”, is a library written in Go and implements foundational principals of OBADA protocol, such as Obit, ObitID, USN and checksum generation.
+
+---
+
+<<< badges here >>>
+
+## Install
+Normally you would simply import "github.com/obada-foundation/sdkgo".
+
+# Examples
