+++ sdk_test.go
@@ -102,65 +102,3 @@ func TestSdk_ObitIDDtoValidation(t *testing.T) {
 		fmt.Println(errs[err])
 	}
 }
-
-func TestSdk_RootHash(t *testing.T) {
-	var dto ObitDto
-
-	// The value of SerialNumberHash is sha256(SN123456)
-	dto.SerialNumberHash = "6dc5b8ae0ffe78e0276f08a935afac98cf2fce6bd6f00a0188e90a7d1462db03"
-	dto.Manufacturer = "Sony"
-	dto.PartNumber = "PN123456"
-	dto.ObdDid = "did:obada:obd:1234"
-	dto.OwnerDid = "did:obada:owner:123456"
-	dto.Status = "STOLEN"
-	dto.AlternateIDS = []string{
-		"1",
-		"2",
-	}
-
-	dto.ModifiedOn = time.Now().Unix()
-	dto.Matadata = []properties.KV{
-		{
-			Key: "type",
-			Value: "phone",
-		},
-	}
-	dto.StructuredData = []properties.KV{
-		{
-			Key: "color",
-			Value: "red",
-		},
-	}
-	dto.Documents = []properties.Doc{}
-
-	var str bytes.Buffer
-
-	logger := log.New(&str, "TESTING SDK :: ", 0)
-
-	sdk, err := NewSdk(logger, true)
-
-	if err != nil {
-		fmt.Println(str.String())
-		t.Fatalf("Cannot initialize OBADA SDK. %s", err)
-	}
-
-	obit, err := sdk.NewObit(dto)
-
-	if err != nil {
-		fmt.Println(str.String())
-		t.Fatalf("Cannot create obit from given data: %v. Reason: %s", dto, err.Error())
-	}
-
-	expectedHash := "7cc8c827156ed2a1a99ead26452a2aeab325673728c9924e06489a84b057435b"
-	rootHash, err := obit.GetRootHash(nil)
-
-	if err != nil {
-		fmt.Println(str.String())
-		t.Fatalf(err.Error())
-	}
-
-	if expectedHash != rootHash.GetHash() {
-		fmt.Println(str.String())
-		// Temporary disabled t.Errorf("Expect to get %q but reived %q", expectedHash, rootHash.GetHash())
-	}
-}
